==> Performance_benchmarks/Bulk_simulations/Hsapiens_dataset/1_Hsapiens_prepare.Rmd <==
---
title: "1_Hsapiens_prepare"
author: "Jeroen Gilis"
date: "06/11/2020"
output: html_document
---

**In order to run this script (1_Hsapiens_prepare.Rmd), the folder Hsapiens_kallisto.zip, which contains the kallisto quantification files for this dataset, should be downloaded from Zenodo and unzipped. In addition, two metaData files Hsapiens_metadata_1.xlsx, Hsapiens_metadata_2.txt should also be downloaded from Zenodo. All three files must then be copied into the Data folder of this Github page.** 

Note that the raw data is available through [ArrayExpress](https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-3766/)

**If one does not want to run this script, its output can also be downloaded from Zenodo: Hsapiens_benchmark_datasets_count.Rdata (or, alternatively, Hsapiens_benchmark_datasets_scaledTPM.Rdata)**

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load libraries

```{r,message=FALSE,warning=FASLE}
library(tximport)
library(DRIMSeq)
library(edgeR)
```

# Load data

```{r}
root.dir <- rprojroot::find_rstudio_root_file()
sample_names <- c("output_01","output_02","output_03","output_04","output_05","output_06","output_07","output_08","output_09","output_10")
quant_files <- file.path(root.dir, "Data/Hsapiens_kallisto", sample_names, "abundance.h5")

txi <- tximport(files = quant_files, type = "kallisto", countsFromAbundance = "no", txOut = TRUE)
#txi <- tximport(files = quant_files, type = "kallisto", countsFromAbundance = "scaledTPM", txOut = TRUE)

myDesign <- data.frame(sample_id = as.character(sample_names),
                       condition = as.character(rep(c("group1","group2")),each=5))

colnames(txi$counts) <- myDesign$sample_id
```

# Load metadata

```{r}
truth <- read.table(file = "./Data/Hsapiens_metadata_1.txt", sep = "\t", header = TRUE)
conversion <- read.table(file = "./Data/Hsapiens_metadata_2.txt")

colnames(conversion) <- c("target_id", "transcript_id")
conversion$target_id <- as.character(conversion$target_id)
conversion$transcript_id <- as.character(conversion$transcript_id)

txInfo <- as.data.frame(cbind(as.character(truth$gene_id), as.character(truth$transcript_id), as.character(truth$transcript_ds_status)))

colnames(txInfo) <- c("gene_id", "transcript_id", "gene_modified")
txInfo$gene_id <- as.character(txInfo$gene_id)
txInfo$transcript_id <- as.character(txInfo$transcript_id)
txInfo <- txInfo[,c(2,1,3)] # change order in same way as other benchmarks
```

# Generate lenient filtering benchmark dataset (edgeR)

```{r}
Hs_lenient <- as.data.frame(txi$counts)
txInfo_lenient <- txInfo

group <- as.factor(rep(c("a","b"),each=5))
design <- model.matrix(~group)
sampleData <- design
sampleData[,1] <- colnames(Hs_lenient)
colnames(sampleData) <- c("sample_id", "condition")

Hs_lenient <- edgeR::DGEList(counts = Hs_lenient,group=group)
filter <- edgeR::filterByExpr(Hs_lenient,design=design)
            
Hs_lenient <- Hs_lenient$counts[filter,]

## Reorder corresponding info
txInfo_lenient <- txInfo_lenient[match(
    rownames(Hs_lenient), txInfo_lenient$transcript_id
),]

## Filter out genes with one TX
txInfo_lenient <- txInfo_lenient[txInfo_lenient$gene_id %in% names(table(txInfo_lenient$gene_id))[table(txInfo_lenient$gene_id) > 1],]

Hs_lenient <- Hs_lenient[txInfo_lenient$transcript_id,]

## add column with sample size for easy retrieval
txInfo_lenient$nrSamplesPerCondition <- 5
colnames(txInfo_lenient) <- c('TXNAME','GENEID','gene_modified','nrSamplesPerCondition')
rownames(txInfo_lenient) <- txInfo_lenient$TXNAME

sampleData <- as.data.frame(sampleData)
sampleData$condition <- as.factor(rep(c("a","b"),each=5))

HsBenchmarkLenient <- list(
            data     = as.data.frame(Hs_lenient),
            design   = sampleData,
            metaInfo = txInfo_lenient
)

dim(HsBenchmarkLenient$data)
```

# Generate stringent filtering benchmark dataset (DRIMSeq)

```{r}
Hs_stringent <- as.data.frame(txi$counts)
txInfo_stringent <- txInfo
group <- as.factor(rep(c("a","b"),each=5))
design <- model.matrix(~group)
sampleData <- design
sampleData[,1] <- colnames(Hs_stringent)
colnames(sampleData) <- c("sample_id", "condition")

Hs_stringent <- Hs_stringent[rownames(Hs_stringent)%in%txInfo$transcript_id,]

geneForEachTx <- txInfo[match(rownames(Hs_stringent),txInfo$transcript_id),"gene_id"]

Hs_stringent$gene_id <- geneForEachTx
Hs_stringent$feature_id <- row.names(Hs_stringent)

d <- DRIMSeq::dmDSdata(counts = Hs_stringent, samples = as.data.frame(sampleData))
d <- dmFilter(d,
              min_samps_feature_expr=5, 
              min_feature_expr=10, 
              min_samps_feature_prop=5, 
              min_feature_prop=0.1,
              min_samps_gene_expr=10, 
              min_gene_expr=10
              )

Hs_stringent <- as.data.frame(txi$counts)[counts(d)$feature_id,]

## Reorder corresponding info
txInfo_stringent <- txInfo_stringent[match(
    rownames(Hs_stringent), txInfo_stringent$transcript_id
),]

## Filter out genes with one TX
txInfo_stringent <- txInfo_stringent[txInfo_stringent$gene_id %in% names(table(txInfo_stringent$gene_id))[table(txInfo_stringent$gene_id) > 1],]

Hs_stringent <- Hs_stringent[txInfo_stringent$transcript_id,]

## add truth column
txInfo_stringent$gene_modified <- truth$transcript_ds_status[match(txInfo_stringent$transcript_id,as.character(truth$transcript_id))]

## add column with sample size for easy retrieval
txInfo_stringent$nrSamplesPerCondition <- 5
colnames(txInfo_stringent) <- c('TXNAME','GENEID','gene_modified','nrSamplesPerCondition')
rownames(txInfo_stringent) <- txInfo_stringent$TXNAME

sampleData <- as.data.frame(sampleData)
sampleData$condition <- as.factor(rep(c("a","b"),each=5))

HsBenchmarkStringent <- list(
            data     = as.data.frame(Hs_stringent),
            design   = as.data.frame(sampleData),
            metaInfo = txInfo_stringent
)

dim(HsBenchmarkStringent$data)
```

# Save Hsapiens benchmark data

```{r}
save(HsBenchmarkLenient, HsBenchmarkStringent, file="./Data/Hsapiens_benchmark_datasets_count.Rdata")
#save(HsBenchmarkLenient, HsBenchmarkStringent, file="./Data/Hsapiens_benchmark_datasets_scaledTPM.Rdata")
```
