==> Performance_benchmarks/Bulk_simulations/Hsapiens_dataset/2_Hsapiens_DTU.Rmd <==
---
title: "2_Hsapiens_DTU"
author: "Jeroen Gilis"
date: "06/11/2020"
output: html_document
---

**In order to run this script (2_Hsapiens_DTU.Rmd), the dataset Hsapiens_benchmark_datasets_count.Rdata (or, alternatively, Hsapiens_benchmark_datasets_scaledTPM.Rdata) is required.** This file can either be generated with the 1_Hsapiens_prepare.Rmd script or downloaded from Zenodo.

Here we run the DTU analyses for 7 DTU methods on both Hsapiens benchmark datasets. This code runs approximately 30 minutes on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DRIMSeq analysis.

**If you do not want to run this script, its output can also be downloaded from Zenodo: Hsapiens_DTU_results_count.Rdata (or, alternatively, Hsapiens_DTU_results_scaledTPM.Rdata)**

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
library(satuRn)
library(doMC)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```{r}
nrCores <- 2

if(nrCores != 1) {
    doParallel <- TRUE
    doProgress <- 'none'

    registerDoMC(cores = nrCores)
} else {
    doParallel <- FALSE
    doProgress <- 'text'
}
```

# Run the analysis for all methods on both Hsapiens datasets

Run the DTU analyses for 7 DTU methods on both Hsapiens benchmark datasets.
This code runs approximately 30 minutes on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime to the analysis with DRIMSeq.

```{r}
### Load benchmark data
load(file="./Data/Hsapiens_benchmark_datasets_count.Rdata")
#load(file="./Data/Hsapiens_benchmark_datasets_scaledTPM.Rdata")

HsBenchmarkLenient <- list(HsBenchmarkLenient)
HsBenchmarkStringent <- list(HsBenchmarkStringent)

names(HsBenchmarkLenient)   <- paste0(names(HsBenchmarkLenient)  , 'filterLenient')
names(HsBenchmarkStringent) <- paste0(names(HsBenchmarkStringent), 'filterStringent')

HsBenchmarkData <- c(HsBenchmarkLenient,HsBenchmarkStringent)

### Run DTU analysis on benchmark data

print("start satuRn")

tStart <- Sys.time()
suppressWarnings(HsDtuBenchmark_satuRn <- plyr::llply(
    .data = HsBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)

print("start edgeR_diffsplice")

tStart <- Sys.time()
HsDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = HsBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DoubleExpSeq")

tStart <- Sys.time()
HsDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = HsBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design,
            quiet=FALSE
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start NBSplice")

tStart <- Sys.time()
HsDtuBenchmark_NBSplice <- plyr::llply(
    .data = HsBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- NBSplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start limma diffsplice")

tStart <- Sys.time()
HsDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = HsBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DEXSeq")

# DEXSeq
tStart <- Sys.time()
HsDtuBenchmark_DEXSeq <- plyr::llply(
    .data = HsBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DRIMSeq")

tStart <- Sys.time()
HsDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = HsBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(HsDtuBenchmark_satuRn) <- paste0('satuRn_', names(HsDtuBenchmark_satuRn))
names(HsDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(HsDtuBenchmark_limmaDiffsplice))
names(HsDtuBenchmark_DEXSeq)  <- paste0('DEXSeq_', names(HsDtuBenchmark_DEXSeq))
names(HsDtuBenchmark_DRIMSeq) <- paste0('DRIMSeq_', names(HsDtuBenchmark_DRIMSeq))
names(HsDtuBenchmark_edgeRdiffsplice) <- paste0('edgeR_diffsplice_', names(HsDtuBenchmark_edgeRdiffsplice))
names(HsDtuBenchmark_NBSplice)  <- paste0('NBSplice_', names(HsDtuBenchmark_NBSplice))
names(HsDtuBenchmark_DoubleExpSeq)  <- paste0('DoubleExpSeq_', names(HsDtuBenchmark_DoubleExpSeq))


### Save result
save(
    HsDtuBenchmark_satuRn,
    HsDtuBenchmark_limmaDiffsplice,
    HsDtuBenchmark_DEXSeq,
    HsDtuBenchmark_DRIMSeq,
    HsDtuBenchmark_edgeRdiffsplice,
    HsDtuBenchmark_NBSplice,
    HsDtuBenchmark_DoubleExpSeq,
    file="./Data/Hsapiens_DTU_results_count.Rdata")

# save(
#     HsDtuBenchmark_satuRn,
#     HsDtuBenchmark_limmaDiffsplice,
#     HsDtuBenchmark_DEXSeq,
#     HsDtuBenchmark_DRIMSeq,
#     HsDtuBenchmark_edgeRdiffsplice,
#     HsDtuBenchmark_NBSplice,
#     HsDtuBenchmark_DoubleExpSeq,
#     file="./Data/Hsapiens_DTU_results_scaledTPM.Rdata")
```
