==> Performance_benchmarks/Bulk_simulations/Hsapiens_dataset/3_Hsapiens_visualize.Rmd <==
---
title: "3_Hsapiens_visualize"
author: "Jeroen Gilis"
date: "08/12/2020"
output: html_document
---

**In order to run this script (3_Hsapiens_visualize.Rmd), the result file of the DTU analysis Hsapiens_DTU_results_count.Rdata (or, alternatively, Hsapiens_DTU_results_scaledTPM.Rdata) is required.** This file can either be generated by running the 2_Hsapiens_DTU.Rmd script or downloaded from Zenodo and put in the data folder of this GitHub repository. **In addition, the ground truth file Hsapiens_metadata_1.txt should also be downloaded from Zenodo and put in the data folder of this GitHub repository.**

The results of this script (FDR-TPR curves) are already available from our GitHub, under the directory ./Results/Hsapiens_benchmark/.

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load libraries

```{r,message=FALSE}
library(ggplot2)
library(iCOBRA)
```

# Load DTU performance results and ground truth file

```{r}
truth <- read.table(file = "./Data/Hsapiens_metadata_1.txt", sep = "\t", header = TRUE)

#load(file="./Data/Hsapiens_DTU_results_count.Rdata")
load(file="./Data/Hsapiens_DTU_results_scaledTPM.Rdata")

HsapiensBenchmark <- c(
    HsDtuBenchmark_satuRn,
    HsDtuBenchmark_limmaDiffsplice,
    HsDtuBenchmark_DEXSeq,
    HsDtuBenchmark_DRIMSeq,
    HsDtuBenchmark_edgeRdiffsplice,
    HsDtuBenchmark_NBSplice,
    HsDtuBenchmark_DoubleExpSeq
)

HsapiensBenchmark  <- HsapiensBenchmark[which(
    sapply(HsapiensBenchmark, function(x) ! is.null(x$dtuAnalysis))
)]

rm(HsDtuBenchmark_satuRn,
    HsDtuBenchmark_limmaDiffsplice,
    HsDtuBenchmark_DEXSeq,
    HsDtuBenchmark_DRIMSeq,
    HsDtuBenchmark_edgeRdiffsplice,
    HsDtuBenchmark_NBSplice,
    HsDtuBenchmark_DoubleExpSeq)
invisible(gc())
```
    
# Helper function 
To get pvalues for the different methods on the different datasets

```{r}
get_pvalues <- function(dataset,nmethods) {
  
  pvalues <- matrix(data=NA, nrow=nrow(dataset[[1]]$dtuAnalysis), ncol=nmethods)
  rownames(pvalues) <- dataset[[1]]$dtuAnalysis$TXNAME
  
  for (i in 1:nmethods) {
    pvalues_new <- data.frame(dataset[[i]]$dtuAnalysis$p_value)
    rownames(pvalues_new) <- dataset[[i]]$dtuAnalysis$TXNAME
    colnames(pvalues_new) <- names(dataset)[i]

    pvalues[,i] <- pvalues_new[match(rownames(pvalues),rownames(pvalues_new)),1]
  }
  
  pvalues <- apply(pvalues, 2, as.numeric)
  rownames(pvalues) <- dataset[[1]]$dtuAnalysis$TXNAME
  pvalues <- as.data.frame(pvalues)
  pvalues <- pvalues[order(rownames(pvalues)),]
  
  colnames(pvalues) <- c("satuRn", "limma_diffsplice", "DEXSeq","DRIMSeq", "edgeR_diffsplice", "NBSplice", "DoubleExpSeq")
  
  return(pvalues)
}
```

# Visualization function

```{r}
visualize_datasets <- function(dataset,nFilters,nmethods) {
  
  print(paste("This function will generate", nFilters, "plots"))

##  We must select the correct dataframes from the list, based on the list names
selection <- c("filterLenient","filterStringent")

## Generate and store the data for the plots
  resList <- c()
  for (j in 1:nFilters) {
      print(j) # print progress
  
      Hsapiens_current <- HsapiensBenchmark[grepl(selection[j], names(HsapiensBenchmark))]
      
      pvalues <- get_pvalues(Hsapiens_current,nmethods)
      
      truth_file_current <- truth[truth$transcript_id %in% rownames(pvalues),]
      truth_file_current <- truth_file_current[order(truth_file_current$transcript_id),]
      rownames(truth_file_current) <- truth_file_current$transcript_id

      cobra <- COBRAData(pval = pvalues, truth = truth_file_current)
      cobra <- calculate_adjp(cobra)
      cobra1perf <- calculate_performance(cobra, binary_truth = "transcript_ds_status", cont_truth = "none", splv = "none", aspects = c("fdrtpr", "fdrtprcurve", "overlap"))

      cobraplot <- prepare_data_for_plot(cobra1perf, colorscheme = "Dark2", facetted = TRUE)
      
      # DRIMSeq 
      new_col <- c(rep(c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "black", "#999999"),4),rep("white",16))
      names(new_col) <- names(cobraplot@plotcolors)
      cobraplot@plotcolors <- new_col
      
      plot_full <- plot_fdrtprcurve(cobraplot, xaxisrange = c(0, 0.35), yaxisrange = c(0,0.85))
    
      #titles <- c("5 versus 5 - edgeR filter - count","5 versus 5 - DRIMSeq filter - count")
      titles <- c("5 versus 5 - edgeR filter - scaledTPM","5 versus 5 - DRIMSeq filter - scaledTPM")
      current_title <- titles[j]

      plot_full <- plot_full + 
        ggtitle(current_title) +
        theme(strip.background = element_blank(),
            strip.text.x = element_blank(),
            plot.title = element_text(size=10),
            axis.text.x = element_text(size = 10),
            axis.text.y = element_text(size = 10),
            axis.title.x = element_text(size = 10),
            axis.title.y = element_text(size = 10))
      
      # png(paste("./Results/Hsapiens_benchmark/FDRTPR_counts_", j, ".png", sep = ""),
      # width     = 5.5,
      # height    = 4.5,
      # units     = "in",
      # res       = 200,
      # pointsize = 4) # start export
      # print(plot_full)
      # dev.off()
      
      png(paste("./Results/Hsapiens_benchmark/FDRTPR_scaledTPM_", j, ".png", sep = ""),
      width     = 5.5,
      height    = 4.5,
      units     = "in",
      res       = 200,
      pointsize = 4) # start export
      print(plot_full)
      dev.off()
  }
  
}
```

```{r}
visualize_datasets(dataset=HsapiensBenchmark,
                   nFilters=2,
                   nmethods=7)
```
